feat: Update default argument value for local collector
This commit updates the default argument value for the local collector in the `build.sh` file. Previously, the default value was set to "no_collector", but now it is set to "yes_collector". This change ensures that the local collector is spawned by default when running the build script, unless specified otherwise.